lista_13_19_20_21

+++ Funcionalidades/system/system.h
@@ -23,7 +23,7 @@ public:
 	void Guardar(string nombreArhivo, Bloque &A_Guardar, size_t Pos);
 	//Falta codear Eliminar, tempalte
 	template<typename T>
-	bool Eliminar(int id, vector<T>&v);
+	bool Eliminar(size_t id, vector<T>&v);
 	vector<Tags> etiquetas(const string& path);
 	bool actualizar_disponibilidad( string nombreArchivo, int id);
